ml_car_price_predection - Copy/requirements.txt
numpy>=1.25
pandas>=2.0
scikit-learn>=1.2
mlflow>=2.6
joblib>=1.2
fastapi>=0.95
uvicorn>=0.22
streamlit>=1.24
shap>=0.42
plotly>=5.0
python-multipart>=0.0.6
psycopg2-binary>=2.9
scipy>=1.10